+++ src/main/java/bumblebee/xchangepass/domain/user/service/UserService.java
@@ -1,7 +1,7 @@
 package bumblebee.xchangepass.domain.user.service;
 
-import bumblebee.xchangepass.domain.user.dto.request.UserRegisterRequest;
 import bumblebee.xchangepass.domain.user.dto.request.UserUpdateRequest;
+import bumblebee.xchangepass.domain.user.dto.response.UserLoginResponse;
 import bumblebee.xchangepass.domain.user.dto.response.UserResponse;
 import bumblebee.xchangepass.domain.user.entity.User;
 import bumblebee.xchangepass.domain.user.repository.UserRepository;
@@ -10,7 +10,6 @@ import bumblebee.xchangepass.global.exception.CommonException;
 import bumblebee.xchangepass.global.util.DuplicateKeyExceptionHandler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,30 +18,6 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserService {
 
     private final UserRepository userRepository;
-    private final NicknameGenerator nicknameGenerator;
-    private final BCryptPasswordEncoder bCryptPasswordEncoder;
-
-    /**
-     * ✅ 사용자 등록
-     * 닉네임 Redis INCR을 활요한 자동 생성
-     * 실명의 경우 추후 전화번호, 이메일에서 받아 오는 형식으로 변경 예정
-     */
-    public void signupUser(UserRegisterRequest request) {
-        String uniqueNickname = null;
-        try{
-            uniqueNickname = nicknameGenerator.generateUniqueNickname();
-            userRepository.save(request.toEntity(bCryptPasswordEncoder, uniqueNickname));
-        } catch (DataIntegrityViolationException e) {
-            nicknameGenerator.rollbackNicknameId(uniqueNickname);
-            DuplicateKeyExceptionHandler.handle(e);
-        } catch (IllegalArgumentException | CommonException e) {
-            nicknameGenerator.rollbackNicknameId(uniqueNickname);
-            throw e;
-        } catch (Exception e) {
-            nicknameGenerator.rollbackNicknameId(uniqueNickname);
-            throw ErrorCode.USER_NOT_REGISTER.commonException();
-        }
-    }
 
     /**
      * ✅ 사용자 조회
@@ -86,4 +61,18 @@ public class UserService {
 
         existUser.softDelete();
     }
+
+    public UserLoginResponse readUserByUserId(String userId) {
+        System.out.println("userId = " + userId);
+        User user = userRepository.findByUserId(Long.parseLong(userId))
+                .orElseThrow(ErrorCode.USER_NOT_FOUND::commonException);
+        System.out.println("user = " + user.toString());
+        return UserLoginResponse.fromEntity(user);
+    }
+
+    public UserLoginResponse readUserByUserEmail(String userEmail) {
+        return UserLoginResponse.fromEntity(userRepository.findByUserEmail(userEmail)
+                .orElseThrow(ErrorCode.USER_NOT_FOUND::commonException));
+    }
+
 }
